+++ Examples/Write-ProgressEx.info.ps1
@@ -1,12 +1,3 @@
-# mazzy@mazzy.ru, 2017-10-21
-# https://github.com/mazzy-ax/write-progressEx
-
-#requires -version 3.0
-
-$module = "Write-ProgressEx.psd1"
-$PSCommandPath | Split-Path -Parent | Split-Path -Parent | Join-Path -ChildPath $module | Import-Module -Force
-
-
 $range = 1..1000
 
 write-ProgressEx 'wait, please' -Total $range.Count