+++ Examples/Write-ProgressEx.parm.ps1
@@ -1,12 +1,3 @@
-# mazzy@mazzy.ru, 2017-10-21
-# https://github.com/mazzy-ax/write-progressEx
-
-#requires -version 3.0
-
-$module = "Write-ProgressEx.psd1"
-$PSCommandPath | Split-Path -Parent | Split-Path -Parent | Join-Path -ChildPath $module | Import-Module -Force
-
-
 $nodes = 1..20
 $names = 1..50